requirements.txt
asgiref==3.8.1
contourpy==1.3.1
crispy-tailwind==1.0.3
cycler==0.12.1
Django==5.1.7
django-admin-interface==0.29.4
django-browser-reload==1.18.0
django-chartjs==2.3.0
django-colorfield==0.12.0
django-crispy-forms==2.3
fonttools==4.56.0
imbalanced-learn==0.13.0
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
sqlparse==0.5.3
text-unidecode==1.3
threadpoolctl==3.5.0
tzdata==2025.1
xgboost==3.0.0